refactor: MineSweeper -> MineSweeperFactory, MineBoard -> MineSweeper 클래스 네이밍 변경
