+++ src/main/kotlin/minesweeper/domain/MineSweeper.kt
@@ -1,22 +1,24 @@
 package minesweeper.domain
 
-data class MineSweeper(val mineCoordinateStrategy: MineCoordinateStrategy = RandomMineCoordinateStrategy) {
+class MineSweeper(
+    val boardFields: BoardFields,
+) {
+    val isEnd get() = boardFields.isAllOpenedNumberFields() || boardFields.isOpenedMineField()
+    val isWin get() = boardFields.isAllOpenedNumberFields()
 
-    fun mineBoard(height: Int, width: Int, mineCount: Int): MineBoard {
-        val coordinates = Coordinate.listOf(height, width)
-        val mineCoordinates = mineCoordinateStrategy.mineCoordinates(coordinates, mineCount)
-        val boardFields = coordinates.map { boardField(mineCoordinates, it) }
-        return MineBoard(BoardFields(boardFields))
+    fun open(coordinate: Coordinate) {
+        require(!isEnd) {
+            "게임이 종료된 후에는 필드를 오픈할 수 없습니다."
+        }
+
+        if (boardFields.open(coordinate) is NumberField) {
+            openNearFields(boardFields.open(coordinate))
+        }
     }
 
-    private fun boardField(
-        mineCoordinates: List<Coordinate>,
-        it: Coordinate
-    ): BoardField {
-        return if (mineCoordinates.contains(it)) {
-            BoardField.mine(it)
-        } else {
-            BoardField.nonMine(it)
-        }
+    private fun openNearFields(field: BoardField) {
+        val notOpenedNumberFields = boardFields.aroundNotOpenedNumberFields(field.coordinate)
+        val openedFields = notOpenedNumberFields.open()
+        openedFields.boardFields.forEach { openNearFields(it) }
     }
 }
